add tests, change database entity

+++ data/common/src/main/java/com/winphyoethu/pocketo/data/common/account/AccountLocalDataSource.kt
@@ -4,8 +4,8 @@ import com.winphyoethu.pocketo.domain.account.model.Account
 
 interface AccountLocalDataSource {
 
-    suspend fun saveAccount(account: Account): Int
+    suspend fun saveAccount(account: Account): Long
 
-    suspend fun updateAccount(account: Account): Int
+    suspend fun updateAccount(account: Account): Long
 
 }
