+++ data/common/src/main/java/com/winphyoethu/pocketo/data/common/expense/ExpenseLocalDataSource.kt
@@ -4,7 +4,7 @@ import com.winphyoethu.pocketo.domain.expense.model.Expense
 
 interface ExpenseLocalDataSource {
 
-    suspend fun saveExpense(expense: Expense) : Int
+    suspend fun saveExpense(expense: Expense) : Long
 
     suspend fun deleteExpense(expense: Expense) : Int
